refactor(core): split finalize phase into dedicated submodules
manager.rs grew to 710 production lines after adding labelled finalize
substeps. The maintainability guardrail asks each Rust source under 600.

Move the phase-substep label registry plus the PhaseProgressEmitter
helper into manager/progress_substep.rs, and the entire Phase 6 body
into manager/finalize.rs as a free function. UpdateManager fields are
crate(super) so finalize.rs can read them without per-field getters.

Behavior unchanged; covered by the existing
test_download_and_apply_persists_current_phase_for_finalize_substeps
and test_download_and_apply_reports_incremental_progress_for_full_update
tests.
